+++ build.mjs
@@ -56,7 +56,7 @@ const ES_BUILD_OPTIONS = {
     "node:console",
     "crypto",
     "uuid",
-    "@llrt/hex",
+    "llrt:hex",
     "os",
     "fs",
     "child_process",
@@ -66,7 +66,7 @@ const ES_BUILD_OPTIONS = {
     "path",
     "events",
     "buffer",
-    "@llrt/xml",
+    "llrt:xml",
     "net",
     "util",
     "url",
@@ -584,7 +584,7 @@ async function buildSdks() {
       plugins: [AWS_SDK_PLUGIN, esbuildShimPlugin([[/^bowser$/]])],
       alias: {
         "@aws-sdk/util-utf8": "@aws-sdk/util-utf8-browser",
-        "fast-xml-parser": "@llrt/xml",
+        "fast-xml-parser": "llrt:xml",
         "@smithy/md5-js": "crypto",
       },
       chunkNames: "llrt-[name]-sdk-[hash]",
