+++ shims/util-hex-encoding.js
@@ -1,4 +1,4 @@
-import { encode, decode } from "@llrt/hex";
+import { encode, decode } from "llrt:hex";
 
 export const fromHex = decode;
 export const toHex = encode;
