+++ tests/unit/encoding.test.ts
@@ -1,6 +1,6 @@
-import hex from "@llrt/hex";
+import hex from "llrt:hex";
 
-describe("@llrt/hex", () => {
+describe("llrt:hex", () => {
   it("should encode/decode text", () => {
     let hello = "hello";
     const encoded = new TextEncoder().encode(hello);
